backend/requirements.txt
flask
flask_sqlalchemy
flask_security_too
flask_restful
python-dotenv